third_party/blink/web_tests/wpt_internal/fenced_frame/navigate-by-name-expected.txt

This is a testharness.js-based test.
FAIL navigate by name assert_equals: The anchor element did not navigate a frame inside the fenced frame boundary: top-level fenced frame expected "PASS" but got "FAIL"
Harness: the test ran to completion.
